initial refactor
updated SamplePlugin to RemotePartyNotifier, updated description and json
