+++ RemotePartyNotifier.sln
