+++ RemotePartyNotifier/Configuration.cs
@@ -2,7 +2,7 @@
 using Dalamud.Plugin;
 using System;
 
-namespace SamplePlugin
+namespace RemotePartyNotifier
 {
     [Serializable]
     public class Configuration : IPluginConfiguration
