+++ RemotePartyNotifier/RemotePartyNotifier.csproj
